pyproject.toml
[tool.poetry]
name = "mictlanxbasics"
version = "0.1.0"
description = ""
authors = ["Ignacio Castillo <ignacio.bcastillo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">3.9,<4"



python-dotenv = "^1.0.1"
requests = "^2.32.3"
mictlanx = "0.0.161a29"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"



[[tool.poetry.source]]
name = "test"
url = "https://test.pypi.org/simple/"
priority = "supplemental"